% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{data.import}
\alias{data.import}
\title{Import genotype data in the correct format for network construction}
\usage{
data.import(ped, tped, gwas_id, gwas_p, pvalue, id.select)
genomic correlations aswell epistatic interactions a genotype matrix has to be
created, consisting of one numeric value per SNP, per individual. This function
takes Plink output (1,2-coding) to create the genotype matrix which can be used
to calculate genomic correlations or epistatic interaction effects
}
\arguments{
\item{ped}{The ped file (.ped) is an input file from Plink: The PED file is a
white-space (space or tab) delimited file: the first six columns are mandatory:
Family ID, Idividual ID, Paternal ID, Maternal ID, 
Sex (1=male; 2=female;other=unknown) and Phenotype. The IDs are alphanumeric: 
the combination of family and individual ID should uniquely identify a person.
A PED file must have1 and only 1 phenotype in the sixth column.
The phenotype can be either a quantitative trait or an affection status 
column: PLINK will automatically detect which type
(i.e. based on whether a value other than 0, 1, 2 or the missing genotype 
code is observed). SNPs are 1,2-coded (1 for major allele,2 for minor allele) 
For more information: http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#ped}

\item{tped}{The tped file (.tped) is a transposed ped file, from Plink. 
This file contains the SNP and genotype information where one row is a SNP.
The first 4 columns of a TPED file are the same as a 4-column MAP file.
Then all genotypes are listed for all individuals for each particular SNP on 
each line. Again, SNPs are 1,2-coded.}

\item{gwas_id}{A vector of all SNPs in the GWAS}

\item{gwas_p}{A vector of the p-values corresponding to the gwas_id vector}

\item{pvalue}{A value for the cutoff of the SNPs which should be remained 
in the matrix, based on the pvalue resulting from the GWAS}

\item{id.select}{If requested, asubset of individuals can be 
selected (e.g. extremes). If nothing inserted, all individuals are in the
output}
}
\value{
A genotype dataframe with a row for each individual and a column
 for each SNP. SNPs are 1,1.5,2 coded: 1 for homozygous for the major 
 allele, 1.5 for heterozygous, and 2 for homozygous for the minor allele. 
 Missing values are NA coded.
}
\description{
For network construction based on both
}
\details{
There is so much to be said
}
\examples{
data.import(ped, tped, gwas_id, gwas_p, pvalue, id.select)


}
\references{
Lisette J.A. Kogelman and Haja N.Kadarmideen (2014). 
Weighted Interaction SNP Hub (WISH) network method for building genetic
networks for complex diseases and traits using whole genome genotype data.
BMC Systems Biology 8(Suppl 2):S5. 
http://www.biomedcentral.com/1752-0509/8/S2/S5.
}

