% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{generate.modules}
\alias{generate.modules}
\title{Visualization of chromosome pairwise region epistatic interaction strength, based on 
statistical significance}
\usage{
generate.modules(correlations)
}
\arguments{
\item{correlations}{List of epistatic correlations and p-values genrated by
epistatic.correlation()}

\item{power}{Powers to test for creating scale free network. Only change if the default
values don't work}

\item{n.snps}{Number of SNPs to select. SNPs are selected by connectivity, so 500 will
select the top 500 most connected Snps. Default is to use all}

\item{minClusterSize}{Minimum module (cluster) size. Default, is 50, but changing this may
be recommended in case of sparse SNPs}

\item{type}{Type of network to generate. Default is "unsigned", can be "signed" or "signed hybrid"}

\item{threads}{Number of threads to use if parallelization is possible.}
}
\value{
Plots the network connectivity and the scale and SNP tree clustering with modules found. 
Returns a named list with all the data generated:
\itemize{
 \item{"SNPs"}{SNPs used in the analysis and their correlations}
 \item{"connectivity"}{The connectivity matrix of the SNPs}
 \item{"adjMat"}{The adjacency matrix of the SNPs}
 \item{"dissTom"}{The dissimilarity TOM}
 \item{"genetree"}{The clustering object used for the genetree}
 \item{"modules"}{The module numbers for each SNP, in order of the SNP matrix}
 \item{"modulcolors"}{The colors used in the modules for each SNP}
 \item{"power.estimate"}{The power estimate to generate a scale free network}
}
}
\description{
Visualization of chromosome pairwise region epistatic interaction strength, based on 
statistical significance. The value is based of the most signficant epistatic interaction in each
region pair, ranging from 1 ( strongest) to 0 (weakest). By defaulty chromosomes are separated into
1 Mb regions, but if SNPs are more spaced out that this it will adjust to the smallest region that fit
the data.
}
\examples{
 generate.modules(correlations)

}

