% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{genome.interaction}
\alias{genome.interaction}
\title{Visualization of pairwise chromosome epistatic interactions on a genome wide level}
\usage{
genome.interaction(tped,correlations)
}
\arguments{
\item{tped}{The tped file used in generate.genotype(). The SNPs must
be sorted by chromosome, matching the order of the SNPs in the correlation 
matrices.}

\item{correlations}{List of epistatic correlations and p-values genrated by
epistatic.correlation()}

\item{quantile}{Number from 0 to 1 indicating which quantile to base the
visualization on.}
}
\value{
Outputs a plot visualizing the chromosome interaction map
}
\description{
Visualization of the genome wide chromosome pairwise relative strength
of epistatic interaction, ranging from 1 (strongest) to -1 (weakest).
The strength is based on the 90th percentile quantile (default) of stastistical
significance of epistatic interaction between all interactions in each
chromosome pair, scaled to 1 to -1.
}
\examples{
 genome.interaction(tped,correlations)

}

