% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{partial_correlations}
\alias{partial_correlations}
\title{This function calculates the epistatic correlations in a subset of 
a matrix space based on coordiantes}
\usage{
partial_correlations(genotype,genotype_rev,phenotype,coords,model)
}
\arguments{
\item{genotype}{Dataframe with the genotype information, resulting from 
the function generate.genotype(). Make sure that the dataframe contains the 
same individuals as in the phenotype-file, and that those are in the 
same order.}

\item{genotype_rev}{Same as genotpye but with reversed genotype coding}

\item{phenotype}{Dataframe with the rows correspinding to the individuals
in the analysis,and columns for the different measured phenotypes and 
fixed/random factors. Phenotypes should be continous variables.}

\item{coords}{Matrix of row split coordinates for subseting input space}

\item{model}{Specification controlling if MM or Mm directed interaction
model is used.}
}
\value{
Epsitatic correlations and P-values for the selected set or subset
of the data
}
\description{
Internal function for calculating epsitatic correlations
in sub-matrices
}
\examples{
partial_correlations <- function(genotype,genotype_rev,phenotype,coords,model)

}

