% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{partial_correlations_triangular}
\alias{partial_correlations_triangular}
\title{***Internal Use Function*** This function calculates the epistatic correlations
in a subset of a matrix space based on coordiantes}
\usage{
partial_correlations_triangular(genotype_1,genotype_rev_1,phenotype,coords,
glm=F,model=1)
}
\arguments{
\item{genotype_1}{Dataframe with the genotype information, resulting from 
the function generate.genotype(). Make sure that the dataframe contains the 
same individuals as in the phenotype-file, and that those are in the 
same order.}

\item{genotype_rev_1}{Same as genotpye but with reversed genotype coding}

\item{phenotype}{Dataframe with the rows correspinding to the individuals
in the analysis,and columns for the different measured phenotypes and 
fixed/random factors. Phenotypes should be continous variables.}

\item{coords}{Matrix of row split coordinates for subseting input space}

\item{glm}{setting controlling if is a lm or glm}

\item{model}{Specification controlling if MM or Mm directed interaction
model is used.}
}
\value{
Epsitatic correlations and P-values for the selected set or subset
of the data
}
\description{
Internal package function for calculating epsitatic correlations
in sub-matrices
}
\examples{
partial_correlations <- partial_correlaiton_triangular(genotype_1,genotype_rev_1,
phenotype,coords,model)

}
