% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{triangular_split}
\alias{triangular_split}
\title{***Internal Use Function*** This function calculates the row coordinates for splitting triangular sub
matrices of quadratic matrices into approximately equally sized partitions
for use in in dividing correlation calculations into equal size for 
parallelization}
\usage{
triangular_split(n, split)
}
\arguments{
\item{n}{Row and Column length of the n by n matrix the triangular matrix
originates from}

\item{split}{Number of partitions to split the triangular matrix in}
}
\value{
A matrix of row coordinates used for splitting
}
\description{
Internal function for splitting triangular matrices into
approximately equal parts
}
\examples{
triangular_split(1000,5)

}

