\name{WISH-package}
\alias{WISH-package}
\alias{WISH}
\docType{package}
\title{
\packageTitle{WISH}
}
\description{
\packageDescription{WISH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WISH}
\packageIndices{WISH}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{WISH}

Maintainer: \packageMaintainer{WISH}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
