% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{epistatic.correlation}
\alias{epistatic.correlation}
\title{Calculate the epistatic interaction effect between SNP pairs  using a genotype 
data frame created from genarate.genotype()}
\usage{
epistatic.correlation(phenotype,genotype,threads=1,test=T,simple=T,glm=F)
}
\arguments{
\item{phenotype}{Dataframe with the rows correspinding to the individuals
in the analysis,and columns for the different measured phenotypes and 
fixed/random factors. Only give one phenotype column at a time. Phenotypes
should be non-categorical continous or discrete/semi-discrete variables. 
Make sure that the dataframe contains the same
individuals as in the genotype-file, and that those are in the same order.}

\item{genotype}{Dataframe with the genotype information, resulting from 
the function generate.genotype(). Make sure that the dataframe contains the 
same individuals as in the phenotype-file, and that those are in the 
same order.}

\item{threads}{Number of threads to use for parallel execution in the function 
registerDoParallel()}

\item{test}{True or False value indicating if a test run is being perform.
If True will calculate the expected time it will take for the full analysis
based on calculating 100.000 models with the setting chosen}

\item{simple}{True or false value indicating if only a major/major and
minor/minor directed interaction model are tested (simple=T) or if if 
interactions on the major/minor minor axis are tested as well, with the 
best one of the two being selected (simple=F).}

\item{glm}{If T will use a generelized linear model with a binomial link
function instead of a regular linear model. This should be used if your
phenotype is binary.}
}
\value{
A list of two matrices. The first matrix gives the epistatic
interaction effects between all the SNP-pairs which were in the input 
genotype data) and selected with the pvalue from the GWAS results. 
The second matrix are the corresponding pvalues of the parameter 
estimates of the epistatic interactions.
}
\description{
A WISH network can be built based on epistatic interaction 
effects between SNP pairs. Those interaction effects are calculated using
linear models.
}
\examples{
epistatic.correlation(phenotype,genotype,threads,test,simple)

}
\references{
Lisette J.A. Kogelman and Haja N.Kadarmideen (2014). 
Weighted Interaction SNP Hub (WISH) network method for building genetic
networks for complex diseases and traits using whole genome genotype data.
BMC Systems Biology 8(Suppl 2):S5. 
http://www.biomedcentral.com/1752-0509/8/S2/S5.
}
