% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{epistatic.correlation}
\alias{epistatic.correlation}
\title{Calculate the epistatic interaction effect between SNP pairs to construct a 
WISH network using a genotype data frame created from genarate.genotype()}
\usage{
epistatic.correlation(phenotype, genotype, parallel)
}
\arguments{
\item{phenotype}{Dataframe with the rows correspinding to the individuals
in the analysis,and columns for the different measured phenotypes and 
fixed/random factors. Only give one phenotype column at a time. Phenotypes
should be continous variables. Make sure that the dataframe contains the same
individuals as in the genotype-file, and that those are in the same order.}

\item{genotype}{Dataframe with the genotype information, resulting from 
the function generate.genotype(). Make sure that the dataframe contains the 
same individuals as in the phenotype-file, and that those are in the 
same order.}

\item{parallel}{Number of cores to use for parallel execution in the function 
registerDoParallel()}
}
\value{
A list of two matrices. The first matrix gives the epistatic
 interaction effects between
all the SNP-pairs which were in the input (genotype data) and selected with
the pvalue from the GWAS results. The second matrix are the corresponding
pvalues of the parameter estimates of the epistatic interactions.
}
\description{
A WISH network can be build based on epistatic interaction 
effects between SNP pairs. Those interaction effects are calculated using
ASReml and can be used directly in the WISH network construction.
}
\examples{
epistatic.correlation(phenotype,genotype,parallel)


}
\references{
Lisette J.A. Kogelman and Haja N.Kadarmideen (2014). 
Weighted Interaction SNP Hub (WISH) network method for building genetic
networks for complex diseases and traits using whole genome genotype data.
BMC Systems Biology 8(Suppl 2):S5. 
http://www.biomedcentral.com/1752-0509/8/S2/S5.
}

