% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{pairwise.chr.map}
\alias{pairwise.chr.map}
\title{Visualization of chromosomal pairwise region epistatic interaction strength, based on 
statistical significance}
\usage{
pairwise.chr.map(chr1,chr2,tped,correlations,span)
}
\arguments{
\item{chr1}{The name of the first chromosome in the comparison, matching the name
from the tped file}

\item{chr2}{The name of the second chromosome in the comparison, matching the name
from the tped file}

\item{tped}{The tped file used in generate.genotype(). The SNPs must
be sorted by chromosome and position on the chromosome, matching the order of the SNPs in the correlation 
matrices.}

\item{correlations}{List of epistatic correlations and p-values genrated by
epistatic.correlation()}

\item{span}{Region in bp. Default is 1 Mb (10^6)}
}
\value{
Outputs a plot visualizing the pairwise chromosome region interaction
}
\description{
Visualization of chromosome pairwise region epistatic interaction strength, based on 
statistical significance. The value is based of the most signficant epistatic interaction in each
region pair, ranging from 1 ( strongest) to 0 (weakest). By defaulty chromosomes are separated into
1 Mb regions, but if SNPs are more spaced out that this it will adjust to the smallest region that fit
the data.
}
\examples{
 pairwise.chr.map("1","2",tped,correlations)

}
