% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{pseudo_manhattan}
\alias{pseudo_manhattan}
\title{Function to plot summary pseudo-manhattan plots of variants.}
\usage{
pseudo_manhattan(tped,correlations,values="p")
}
\arguments{
\item{tped}{Input tped file as .tped file or data frame. The tped file (.tped)
is a transposed ped file, from Plink. 
This file contains the SNP and genotype information where one row is a SNP.
The first 4 columns of a TPED file are the same as a 4-column MAP file.
Then all genotypes are listed for all individuals for each particular SNP on 
each line. Again, SNPs are 1,2-coded.#'}

\item{correlations}{List of epistatic correlations and p-values generated by
epistatic.correlation()}

\item{values}{Values can either be p or for p-values or c for coefficients
depending on if you wan't to use the effects sizes for p-values for plotting}
}
\value{
Plots a pseudo manhattan plot
}
\description{
Visualize summary statistics for interactions based on total sum 
of -loglikelihoods for each variant across all interactions og the sum of
effect sizes.
}
\examples{
pseudo_manhattan(tped,correlations)

}
