\name{WISH-package}
\alias{WISH-package}
\alias{WISH}
\docType{package}
\title{
\packageTitle{WISH}
}
\description{
The Weighted Interaction SNP Hub network method uses high-throughput genotype data to detect genome-wide interactions between SNPs and its relation with complex traits. Data dimensionality reduction is achieved by selecting SNPs based on its degree of genome-wide significance and degree of genetic variation in a population. Network construction is based on the epistatic interaction effect between SNP pairs. To identify modules the Topological Overlap Measure is calculated, reflecting the degree of overlap in shared neighbours between SNP pairs. Modules, clusters of highly interconnected SNPs, are defined using a tree-cutting algorithm on the SNP dendrogram created from the dissimilarity TOM.
}
\author{
Victor A.O Caremlo, Lisette Kogelamn and Haja Kadarmideen
Maintainer: hajak@sund.ku.dk
}
\references{
 Lisette J.A. Kogelman and Haja N.Kadarmideen (2014). 
Weighted Interaction SNP Hub (WISH) network method for building genetic 
Networks for complex diseases and traits using whole genome genotype data. 
BMC Systems Biology 8(Suppl 2):S5.  
http://www.biomedcentral.com/1752-0509/8/S2/S5.
}

